+++ src/main/scala/com/atomist/rug/spi/InstantEditorFailureException.scala
@@ -1,8 +1,0 @@
-package com.atomist.rug.spi
-
-/**
-  * Thrown when an editor should fail instantly
-  *
-  * @param msg message explaining failure
-  */
-class InstantEditorFailureException(msg: String) extends RuntimeException(msg)